fix(#4058): contain MemBackend to cfg(test) — drop leaky test-utils feature
Remove the `test-utils` Cargo feature from kernel so feature unification
can no longer leak `MemBackend`/`with_mem_backend` into production builds
when transport lists kernel in [dev-dependencies]. Restrict the kernel-side
helpers to `#[cfg(test)]` (kernel-internal only). Replace the transport
grpc.rs test dependency on `Kernel::with_mem_backend()` with an in-tree
`MemBackend` struct (also `#[cfg(test)]`) wired via the public
`Kernel::sys_setattr(DT_MOUNT)` path — no cross-crate test feature needed.
