+++ rust/transport/Cargo.toml
@@ -51,5 +51,3 @@ pyo3 = { version = "0.28.3", features = ["hashbrown"], optional = true }
 [dev-dependencies]
 tempfile = "3"
 rcgen = { version = "0.14", features = ["pem", "x509-parser"] }
-# `test-utils` exposes `Kernel::with_mem_backend()` for grpc integration tests.
-kernel = { workspace = true, features = ["test-utils"] }
